requirements.txt
whois
requests
dnspython
python-dotenv
pygments
validators